daos implemented

+++ src/com/git/programmerr47/vkazam/daos/FingerprintDao.java
@@ -18,11 +18,11 @@ public class FingerprintDao {
         databaseHelper = new DatabaseHelper(context);
     }
 
-    public Fingerprint get(int id) {
+    public Fingerprint get(long id) {
         SQLiteDatabase db = databaseHelper.getReadableDatabase();
         Cursor cursor = db.query(FingerprintTable.FINGERPRINTS, null, FingerprintTable._ID + "=?",
                 new String[] {
-                    Integer.toString(id)
+                    Long.toString(id)
                 }, null, null, null);
         Fingerprint fingerprint = null;
         if (cursor.moveToFirst()) {
@@ -67,18 +67,18 @@ public class FingerprintDao {
             result = db.update(FingerprintTable.FINGERPRINTS, values, FingerprintTable._ID
                     + "=?",
                     new String[] {
-                        Integer.toString(fingerprint.getId())
+                        Long.toString(fingerprint.getId())
                     });
         }
         db.close();
         return result;
     }
 
-    public int delete(int id) {
+    public int delete(long id) {
         SQLiteDatabase db = databaseHelper.getWritableDatabase();
         int result = db.delete(FingerprintTable.FINGERPRINTS, FingerprintTable._ID + "=?",
                 new String[] {
-                    Integer.toString(id)
+                    Long.toString(id)
                 });
         db.close();
         return result;
